optimize parquet loading to read files one by one

+++ vectordb_bench/backend/clients/api.py
@@ -231,62 +231,49 @@ class PgDB(VectorDB):
         # it handles buffering, matching globs, etc.
         log.info(f"loading {len(parquet_files)} files")
         log.info(f"loading files: {parquet_files}")
-        dataset = ds.dataset(parquet_files)
-
-        # create an encoder object which will do the encoding
-        # and give us the expected Postgres table schema
-        encoder = ArrowToPostgresBinaryEncoder(dataset.schema)
-        # get the expected Postgres destination schema
-        # note that this is _not_ the same as the incoming arrow schema
-        # and not necessarily the schema of your permanent table
-        # instead it's the schema of the data that will be sent over the wire
-        # which for example does not have timezones on any timestamps
-
-        # here, since embedding tables just have id and float4[] columns, 
-        # it is fine to actualy use this table as the final table
-        pg_schema = encoder.schema()
-        # assemble ddl for a temporary table
-        # it's often a good idea to bulk load into a temp table to:
-        # (1) Avoid indexes
-        # (2) Stay in-memory as long as possible
-        # (3) Be more flexible with types
-        #     (you can't load a SMALLINT into a BIGINT column without casting)
 
         tmp_table_name = "_tmp_parquet_data"
-        typed_cols = [f'"{col_name}" {col.data_type.ddl()}' for col_name, col in pg_schema.columns]
-        cols = [col_name for col_name, _ in pg_schema.columns]
-        cols_joined = ','.join(cols)
-        typed_cols_joined = ','.join(typed_cols)
-        self.pg_session.execute(f'DROP TABLE IF EXISTS "{tmp_table_name}"')
-        ddl = f"CREATE TEMPORARY TABLE {tmp_table_name} ({typed_cols_joined})"
-        self.pg_session.execute(ddl)
-
-        log.debug(f"pg schema {pg_schema}")
-        log.debug(f"Assuming underlying postgres table was created with columns: {typed_cols}")
 
         count = 0
         i = 0
-        for batch in dataset.to_batches():
+
+        for file in parquet_files:
+            log.debug(f"Loading file {file}")
+            dataset = ds.dataset(file)
             encoder = ArrowToPostgresBinaryEncoder(dataset.schema)
-            log.info(f"batch: {i} batch len: {len(batch)}")
-            # Write to buffer
-            write_start = time.time()
-            self.binary_f.truncate(0)
-            self.binary_f.seek(0)
-            self.binary_f.write(encoder.write_header())
-            self.binary_f.write(encoder.write_batch(batch))
-            self.binary_f.write(encoder.finish())
-            log.debug(f"Writing batch to buffer took {int(time.time() - write_start)}s")
+            if i == 0:
+                pg_schema = encoder.schema()
+                typed_cols = [f'"{col_name}" {col.data_type.ddl()}' for col_name, col in pg_schema.columns]
+                cols = [col_name for col_name, _ in pg_schema.columns]
+                cols_joined = ','.join(cols)
+                typed_cols_joined = ','.join(typed_cols)
+                self.pg_session.execute(f'DROP TABLE IF EXISTS "{tmp_table_name}"')
+                ddl = f"CREATE TEMPORARY TABLE {tmp_table_name} ({typed_cols_joined})"
+                log.debug(f"pg schema {pg_schema}")
+                log.debug(f"Assuming underlying postgres table was created with columns: {typed_cols}")
+                self.pg_session.execute(ddl)
 
-            # Copy to tmp table
-            copy_start = time.time()
-            self.binary_f.seek(0)
-            self.pg_session.copy_expert(f'COPY "{tmp_table_name}" ({cols_joined}) FROM STDIN WITH (FORMAT BINARY)', self.binary_f)
-            log.debug(f"Writing batch to postgres took {int(time.time() - copy_start)}s, total batch processing took: {int(time.time() - write_start)}s")
-            count += len(batch)
-            i += 1
+            for batch in dataset.to_batches():
+                encoder = ArrowToPostgresBinaryEncoder(dataset.schema)
+                log.info(f"batch: {i} batch len: {len(batch)}")
+                # Write to buffer
+                write_start = time.time()
+                self.binary_f.truncate(0)
+                self.binary_f.seek(0)
+                self.binary_f.write(encoder.write_header())
+                self.binary_f.write(encoder.write_batch(batch))
+                self.binary_f.write(encoder.finish())
+                log.debug(f"Writing batch to buffer took {int(time.time() - write_start)}s")
 
-        # clean buffer
+                # Copy to tmp table
+                copy_start = time.time()
+                self.binary_f.seek(0)
+                self.pg_session.copy_expert(f'COPY "{tmp_table_name}" ({cols_joined}) FROM STDIN WITH (FORMAT BINARY)', self.binary_f)
+                log.debug(f"Writing batch to postgres took {int(time.time() - copy_start)}s, total batch processing took: {int(time.time() - write_start)}s")
+                count += len(batch)
+                i += 1
+
+        # clear buffer
         self.binary_f.seek(0)
         self.binary_f.truncate(0)
